Split CancelOrSaveForLaterButtonComponent and render when authorized
